Create HomePage component
