+++ react/AppClean/package.json
@@ -13,6 +13,9 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.74.5",
+    "react-native-paper": "^5.12.5",
+    "react-native-safe-area-context": "4.10.5",
+    "react-native-vector-icons": "^10.2.0",
     "react-native-web": "~0.19.10"
   },
   "devDependencies": {
